Fix parsing Mediatailor DASH manifests
When thumbnails are present in the stream, MediaTailor formats presentationTimeOffset with a float value,
that made ExoPlayer to fail to parse the manifest. This commit ensures decoding of Long and Int numbers
in float format before converting to their respective primitive values.

+++ library/dash/src/main/java/com/google/android/exoplayer2/source/dash/manifest/DashManifestParser.java
@@ -1754,12 +1754,12 @@ public class DashManifestParser extends DefaultHandler
 
   protected static int parseInt(XmlPullParser xpp, String name, int defaultValue) {
     String value = xpp.getAttributeValue(null, name);
-    return value == null ? defaultValue : Integer.parseInt(value);
+    return value == null ? defaultValue : (int) Double.parseDouble(value);
   }
 
   protected static long parseLong(XmlPullParser xpp, String name, long defaultValue) {
     String value = xpp.getAttributeValue(null, name);
-    return value == null ? defaultValue : Long.parseLong(value);
+    return value == null ? defaultValue : (long) Double.parseDouble(value);
   }
 
   protected static float parseFloat(XmlPullParser xpp, String name, float defaultValue) {
